add convert date utils; update order and structure of course page
